feat(airtable-migration): refactorings - extract import methods - move fieldParser into utils - add documentID radio buttons
